+++ CHANGES.rst
@@ -45,10 +45,6 @@ Changes:
 
 Fixes:
 ------
-- Fix job percent progress reported in logs to be more consistent with actual execution of the process
-  (fixes `#90 <https://github.com/crim-ca/weaver/issues/90>`_).
-- Fix `Job` duration not stopped incrementing when its execution failed due to raised error
-  (fixes `#222 <https://github.com/crim-ca/weaver/issues/222>`_).
 - Fix ``weaver.config.get_weaver_config_file`` called with empty path to be resolved just as requesting the default
   file path explicitly instead of returning an invalid directory.
 - Fix `CWL` package path resolution under Windows incorrectly parsed partition as URL protocol.
@@ -57,6 +53,22 @@ Fixes:
 - Fix potential invalid database state that could have saved an invalid process although the following
   ``ProcessSummary`` schema validation would fail and return ``HTTPBadRequest [400]``. The process is now saved only
   after complete and successful schema validation.
+
+`2.2.0 <https://github.com/crim-ca/weaver/tree/2.2.0>`_ (2021-03-03)
+========================================================================
+
+Changes:
+--------
+- Add ``weaver.wps.utils.get_wps_client`` function to handle the creation of ``owslib.wps.WebProcessingService`` client
+  with appropriate request options configuration from application settings.
+
+Fixes:
+------
+- Fix job percent progress reported in logs to be more consistent with actual execution of the process
+  (fixes `#90 <https://github.com/crim-ca/weaver/issues/90>`_).
+- Fix `Job` duration not stopped incrementing when its execution failed due to raised error
+  (fixes `#222 <https://github.com/crim-ca/weaver/issues/222>`_).
+- Improve race condition handling of ``builtin`` process registration at application startup.
 
 `2.1.0 <https://github.com/crim-ca/weaver/tree/2.1.0>`_ (2021-02-26)
 ========================================================================